# src/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Projeto Trainee Backend - API",
    "description": "Projeto para capacitação trainee de backend, com o objetivo de criar uma API para gerenciar projetos e tarefas."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearer_token": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "name": "bearer_token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/project": {
      "get": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para retornar todos os projetos não concluídos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Projetos encontrados."
          },
          "400": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para cadastrar um projeto",
        "parameters": [
          {
            "name": "name, description, client, deadline, tasks, squad",
            "description": "Dados do projeto.",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "client": {
                  "example": "any"
                },
                "deadline": {
                  "example": "any"
                },
                "tasks": {
                  "example": "any"
                },
                "squad": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto e pastas criado com sucesso."
          },
          "400": {
            "description": "Erro ao criar projeto e pastas."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/project/{projectId}": {
      "get": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para retornar um projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto encontrado."
          },
          "400": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para atualizar um projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "client": {
                  "example": "any"
                },
                "deadline": {
                  "example": "any"
                },
                "tasks": {
                  "example": "any"
                },
                "squad": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto alterado com sucesso."
          },
          "400": {
            "description": "Erro ao atualizar projeto."
          },
          "404": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para excluir um projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto deletado com sucesso."
          },
          "400": {
            "description": "Erro remover projeto."
          },
          "404": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/project/featured-image/{projectId}": {
      "post": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para cadastrar uma imagem de destaque do projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem enviada com sucesso."
          },
          "400": {
            "description": "Erro ao enviar imagem."
          },
          "500": {
            "description": "Erro ao enviar imagem."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para excluir a imagem de destaque do projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Imagem deletada com sucesso."
          },
          "400": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/project/files/{projectId}": {
      "post": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para cadastrar arquivos do projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo enviado com sucesso."
          },
          "404": {
            "description": "Erro ao encontrar projeto."
          },
          "500": {
            "description": "Erro ao enviar arquivo."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/project/complete/{projectId}": {
      "put": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para concluir um projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto concluído com sucesso."
          },
          "400": {
            "description": "Erro ao atualizar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/project/{projectId}/{fileName}": {
      "delete": {
        "tags": [
          "Project"
        ],
        "description": "Endpoint para excluir um arquivo do projeto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do projeto."
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Nome do arquivo."
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo deletado com sucesso."
          },
          "400": {
            "description": "Erro ao encontrar projeto."
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
...
}